requirements.txt
pip~=20.1
lxml~=4.5.0
setuptools~=41.2.0
requests~=2.23.0
boto3~=1.13.1
redis~=3.5.0
PyYAML~=5.3.1
botocore~=1.16.1
Flask~=1.1.1